+++ preload.js
@@ -160,6 +160,7 @@ const api = {
     getSessionCashSummary: (data) => ipcRenderer.invoke('cash:getSessionCashSummary', data),
     getSessionReport: (data) => ipcRenderer.invoke('cash:getSessionReport', data),
     getIncomeReceipts: (data) => ipcRenderer.invoke('cash:getIncomeReceipts', data),
+    searchIncomeReceipts: (data) => ipcRenderer.invoke('cash:searchIncomeReceipts', data),
     createIncomeReceipt: (data) => ipcRenderer.invoke('cash:createIncomeReceipt', data),
     cancelIncomeReceipt: (data) => ipcRenderer.invoke('cash:cancelIncomeReceipt', data),
   },